feat: ajustando código para nova forma de criar pasta, pegando dados de temperatura, e finalizando medição ao chegar na última medição daquela pasta

+++ VM/simulation.py
@@ -5,8 +5,7 @@ import sys
 import time
 
 from VM.FEMM_Simulation import FEMMSimulationController
-from app_service import pegar_ultimo_dado_do_sensor, enviar_pasta_dos_resultados_simulacao, pegar_data_formatada, \
-    pegar_hora_formatada, second_to_hour_minute
+from app_service import *
 from VM.log_handler import write_log
 
 
@@ -34,7 +33,7 @@ class SimulationController:
             "4": {"done": False, "range": 4.3, "in_queue": False},
             "actual_simulation": 0,
             "index_queue": 0,
-            "folder_name": ""
+            "folder_name": self.simulations["folder_name"]
         }
         self.femm.reset()
         if os.path.exists("femm_generate_files") is False:
@@ -79,13 +78,17 @@ class SimulationController:
             with open("simulation.json", "w") as file:
                 file.write(json.dumps(self.simulations, indent=1))
 
-    def update_queue(self):
+    def update_queue(self, folder_name):
+        self.simulations["folder_name"] = folder_name
         self.simulations[str(self.simulations["index_queue"])]["in_queue"] = True
         self.simulations["index_queue"] += 1
         with open("simulation.json", "w") as file:
             file.write(json.dumps(self.simulations, indent=1))
 
-    def start_simulation(self, range, sensor):
+    def is_folder_exist(self):
+        return self.simulations["folder_name"] in os.listdir("./")
+
+    def start_simulation(self, range_sensor, sensor):
         start = time.time()
         try:
             self.load_simulation()
@@ -94,21 +97,20 @@ class SimulationController:
             print("Exception in load_simulation or reset_config: ", e.args)
 
 
-        if (self.is_can_start_simulation(range) is False and sensor.lower() != "sensora"):  # sai da simulaçao e define sensor
+        if (self.is_can_start_simulation(range_sensor) is False and sensor.lower() != "sensora"):  # sai da simulaçao e define sensor
             # Isso aqui serve pra ele iniciar a simulação apenas com o sensorA.
             # Devido a isso, todos os outros dados devem ser setados anteriormente.
             # SensorA deve ser enviado por último.
             return
 
+        write_log(f"!!Nome da pasta atual: {self.simulations['folder_name']}")
         try:
-            if self.simulations["actual_simulation"] == 0:
-                new_folder = self.hash_generate()
-                os.mkdir(new_folder)
-                self.simulations["folder_name"] = new_folder
+            if self.simulations["actual_simulation"] == 0 and self.is_folder_exist() is False:
+                os.mkdir(self.simulations["folder_name"])
                 write_log(f"-Criou pasta com o nome: {self.simulations['folder_name']}\n", folder=self.simulations["folder_name"], create=True)
 
                 with open("folder_name.txt", "w") as file:
-                    file.write(new_folder)
+                    file.write(self.simulations["folder_name"])
         except Exception as e:
             msg = f"Exception in create folder: {e.args}"
             print(msg)
@@ -120,16 +122,17 @@ class SimulationController:
             write_log(f"# INICIO DA SIMULAÇÃO {self.simulations['actual_simulation']}: {pegar_data_formatada()}_{pegar_hora_formatada()}\n", folder=self.simulations["folder_name"])
             sensorB = float(pegar_ultimo_dado_do_sensor("SensorB")["medicao"])
             sensorC = float(pegar_ultimo_dado_do_sensor("SensorC")["medicao"])
+            temperatura = float(pegar_ultimo_dado_do_sensor("SensorTemp")["medicao"])
             write_log(
-                f"# MEDICAO_A: {range} | B: {sensorB} | C: {sensorC}")
-            self.femm.set_femm_atributes(ca=range, cb=sensorB, cc=sensorC,
+                f"# MEDICAO_A: {range_sensor} | B: {sensorB} | C: {sensorC}")
+            self.femm.set_femm_atributes(ca=range_sensor, cb=sensorB, cc=sensorC, temperatura=temperatura,
                                          index=self.simulations["actual_simulation"],
                                          first=self.simulations["actual_simulation"] == 0,
                                          folder_name=self.simulations["folder_name"])
         except Exception as e:
             msg = f"Exception in get SensorB/C and set femm attributes: {e.args}"
             print(msg)
-            write_log(f"\n# MEDICAO_A: {range} | B: {sensorB} | C: {sensorC}\n{msg}")
+            write_log(f"\n# MEDICAO_A: {range_sensor} | B: {sensorB} | C: {sensorC}\n{msg}")
 
 
         try:
@@ -152,6 +155,11 @@ class SimulationController:
             write_log(f"-arquivo simulation.json: {json.dumps(self.simulations)}\n")
             finish = time.time()
             write_log(f"# FIM DA SIMULACAO: {pegar_data_formatada()}_{pegar_hora_formatada()} - TEMPO DE EXECUÇÃO: {second_to_hour_minute(finish - start)}\n")
+
+            if self.simulations["actual_simulation"] == 4:
+                definir_simulacao_completed()
+                write_log(f"\n# FINALIZOU AS SIMULAÇÕES PARA A PASTA: {self.simulations['folder_name']}")
+
         except Exception as e:
             msg = f"Exception in finalize: {e.args}"
             print(msg)
